requirements.txt
autopep8==1.5
azure-devops==6.0.0b2
azure-functions==1.2.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-api-core==1.17.0
google-auth==1.14.1
googleapis-common-protos==1.51.0
idna==2.9
isodate==0.6.0
msrest==0.6.13
oauthlib==3.1.0
opencensus==0.7.7
opencensus-context==0.1.1
opencensus-ext-azure==1.0.2
pika==1.1.0
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
